tec(104324): Refatora arquivo dre/tasks.py
