+++ sme_ptrf_apps/dre/api/views/relatorios_consolidados_dre_viewset.py
@@ -698,14 +698,6 @@ class RelatoriosConsolidadosDREViewSet(GenericViewSet):
             return Response(erro, status=status.HTTP_400_BAD_REQUEST)
 
         try:
-            # gerar_lauda_csv_async.delay(
-            #     dre_uuid,
-            #     tipo_conta_uuid,
-            #     periodo_uuid,
-            #     dados['parcial'],
-            #     request.user.username
-            # )
-
             gerar_lauda_txt_async.delay(
                 dre_uuid,
                 tipo_conta_uuid,
@@ -713,7 +705,6 @@ class RelatoriosConsolidadosDREViewSet(GenericViewSet):
                 dados['parcial'],
                 request.user.username
             )
-
 
         except Exception as err:
             erro = {
