+++ sme_ptrf_apps/dre/tasks/__init__.py
@@ -1,0 +1,8 @@
+from .concluir_consolidado_de_publicacoes_parciais import concluir_consolidado_de_publicacoes_parciais_async
+from .concluir_consolidado_dre import concluir_consolidado_dre_async
+from .gerar_arquivo_ata_parecer_tecnico import gerar_arquivo_ata_parecer_tecnico_async
+from .gerar_previa_consolidado_dre import gerar_previa_consolidado_dre_async
+from .gerar_lauda_txt import gerar_lauda_txt_async
+from .atualiza_regularidade_em_massa import atualiza_regularidade_em_massa_async
+from .gerar_notificacao_prazo_para_acerto_consolidado_dre_apos_vencimento import gerar_notificacao_prazo_para_acerto_consolidado_dre_apos_vencimento_async
+from .gerar_notificacao_prazo_para_acerto_antes_vencimento import gerar_notificacao_prazo_para_acerto_antes_vencimento_async
