+++ sme_ptrf_apps/dre/tasks/concluir_consolidado_dre.py
@@ -1,0 +1,228 @@
+import logging
+from django.db.models import Q
+
+from celery import shared_task
+
+from sme_ptrf_apps.core.models import (
+    Periodo,
+    Unidade,
+    TipoConta,
+    PrestacaoConta
+)
+
+from sme_ptrf_apps.dre.models import (
+    AtaParecerTecnico,
+    ConsolidadoDRE,
+    Lauda,
+    JustificativaRelatorioConsolidadoDRE
+)
+
+from django.core.exceptions import ValidationError
+from django.contrib.auth import get_user_model
+
+logger = logging.getLogger(__name__)
+
+
+def verificar_se_gera_ata_parecer_tecnico(dre=None, periodo=None, consolidado_dre=None, usuario=None, ata=None,
+                                          parcial=None):
+    from .gerar_arquivo_ata_parecer_tecnico import gerar_arquivo_ata_parecer_tecnico_async
+
+    logger.info(f'Iniciando a verificação para gerar a Ata de Parecer Técnico')
+
+    dre_uuid = dre.uuid
+    periodo_uuid = periodo.uuid
+
+    if ata and ata.preenchida_em:
+        logger.info(f'Atrelando a Ata de Parecer Técnico {ata} ao Consolidado DRE {consolidado_dre}')
+        ata.atrelar_consolidado_dre(consolidado_dre)
+        logger.info(
+            f'Iniciando a geração do Arquivo da Ata de Parecer Técnico da DRE {dre}, Período {periodo} '
+            f'e Consolidado DRE {consolidado_dre}')
+        ata_uuid = ata.uuid
+        gerar_arquivo_ata_parecer_tecnico_async(ata_uuid, dre_uuid, periodo_uuid, usuario, parcial)
+    else:
+        logger.info("Ata não preenchida, portanto Arquivo PDF não será gerado")
+
+
+def gerar_lauda_txt_consolidado_dre(consolidado_dre, dre, periodo, parcial, usuario, apenas_nao_publicadas):
+    logger.info(
+        f'Iniciando a geração do arquivo txt da lauda async. DRE:{dre} Período:{periodo} Consolidado DRE {consolidado_dre}.')
+    from sme_ptrf_apps.dre.services import gerar_arquivo_lauda_txt_consolidado_dre
+
+    usuario = get_user_model().objects.get(username=usuario)
+
+    lauda = Lauda.criar_ou_retornar_lauda(
+        consolidado_dre=consolidado_dre,
+        dre=dre,
+        periodo=periodo,
+        usuario=usuario,
+    )
+
+    lauda.passar_para_status_em_processamento()
+
+    logger.info(f"Objeto {lauda} criado/retornado com sucesso")
+
+    try:
+        ata = AtaParecerTecnico.objects.filter(consolidado_dre=consolidado_dre, dre=dre, periodo=periodo).last()
+        if not ata:
+            raise ValidationError(f"O objeto Ata para a DRE {dre} e Período {periodo} não foi encontrado na base.")
+    except (AtaParecerTecnico.DoesNotExist, ValidationError):
+        erro = {
+            'erro': 'Objeto não encontrado.',
+            'mensagem': f"O objeto ata parecer tecnico para a dre {dre.uuid} e periodo {periodo.uuid} não foi encontrado na base."
+        }
+        logger.error('Erro: %r', erro)
+        raise Exception(erro)
+
+    try:
+        nome_dre = dre.nome.upper()
+        if "DIRETORIA REGIONAL DE EDUCACAO" in nome_dre:
+            nome_dre = nome_dre.replace("DIRETORIA REGIONAL DE EDUCACAO", "")
+            nome_dre = nome_dre.strip()
+
+        nome_dre = nome_dre.lower()
+        gerar_arquivo_lauda_txt_consolidado_dre(lauda, dre, periodo, ata, nome_dre, parcial, apenas_nao_publicadas)
+    except Exception as err:
+        erro = {
+            'erro': 'problema_geracao_txt',
+            'mensagem': 'Erro ao gerar txt.'
+        }
+        logger.error("Erro ao gerar lauda: %s", str(err))
+        raise Exception(erro)
+
+    logger.info(
+        f'Finalizado geração arquivo txt da lauda async. DRE:{dre} Período:{periodo}.')
+
+
+def passar_pcs_do_relatorio_para_publicadas(dre, periodo, consolidado_dre):
+    dre_uuid = dre.uuid
+    periodo_uuid = periodo.uuid
+
+    prestacoes = PrestacaoConta.objects.filter(periodo__uuid=periodo_uuid, associacao__unidade__dre__uuid=dre_uuid)
+    prestacoes = prestacoes.filter(Q(status='APROVADA') | Q(status='APROVADA_RESSALVA') | Q(status='REPROVADA'))
+
+    if consolidado_dre.eh_retificacao:
+        # Retificacoes precisam filrar as PCs pelo consolidado vinculado
+
+        prestacoes = prestacoes.filter(publicada=False, consolidado_dre=consolidado_dre)
+    else:
+        prestacoes = prestacoes.filter(publicada=False)
+
+    for prestacao in prestacoes:
+        logger.info(f'Passando Prestação de Contas para Publicada: Prestação {prestacao}')
+        prestacao.passar_para_publicada()
+
+
+def verifica_se_relatorio_consolidado_deve_ser_gerado(dre, periodo, usuario):
+    from .concluir_consolidado_de_publicacoes_parciais import concluir_consolidado_de_publicacoes_parciais_async
+
+    qtde_unidades_na_dre = Unidade.objects.filter(
+        dre_id=dre,
+    ).exclude(associacoes__cnpj__exact='').count()
+
+    qtde_pcs_publicadas_no_periodo_pela_dre = PrestacaoConta.objects.filter(
+        periodo=periodo,
+        status__in=['APROVADA', 'APROVADA_RESSALVA', 'REPROVADA'],
+        associacao__unidade__dre=dre,
+        publicada=True
+    ).count()
+
+    if (int(qtde_unidades_na_dre) == int(qtde_pcs_publicadas_no_periodo_pela_dre)):
+        concluir_consolidado_de_publicacoes_parciais_async(dre.uuid, periodo.uuid, usuario)
+
+
+@shared_task(
+    retry_backoff=2,
+    retry_kwargs={'max_retries': 8},
+    time_limit=333333,
+    soft_time_limit=333333
+)
+def concluir_consolidado_dre_async(
+    dre_uuid=None,
+    periodo_uuid=None,
+    parcial=None,
+    usuario=None,
+    consolidado_dre_uuid=None,
+    ata_uuid=None,
+    sequencia_de_publicacao=None,
+    apenas_nao_publicadas=False,
+):
+    from ..services.consolidado_dre_service import atrelar_pc_ao_consolidado_dre, gerar_relatorio_consolidado_dre
+
+    tipo_contas = TipoConta.objects.all()
+
+    dre = Unidade.dres.get(uuid=dre_uuid)
+    periodo = Periodo.objects.get(uuid=periodo_uuid)
+    ata = AtaParecerTecnico.objects.get(uuid=ata_uuid)
+
+    if consolidado_dre_uuid:
+        consolidado_dre = ConsolidadoDRE.by_uuid(consolidado_dre_uuid)
+    else:
+        consolidado_dre = ConsolidadoDRE.objects.get(dre=dre, periodo=periodo,
+                                                     sequencia_de_publicacao=sequencia_de_publicacao)
+
+    if not consolidado_dre.eh_retificacao:
+        # Uma retificacao ja possui suas PCs vinculadas
+
+        atrelar_pc_ao_consolidado_dre(
+            dre=dre,
+            periodo=periodo,
+            consolidado_dre=consolidado_dre,
+        )
+
+    # Atrelando consolidado as justificativas
+    for tipo_conta in tipo_contas:
+        justificativa = JustificativaRelatorioConsolidadoDRE.objects.filter(
+            dre=dre,
+            tipo_conta=tipo_conta,
+            periodo=periodo,
+            consolidado_dre__isnull=True,
+        ).last()
+
+        if justificativa:
+            justificativa.consolidado_dre = consolidado_dre
+            justificativa.save()
+
+    gerar_relatorio_consolidado_dre(
+        dre_uuid=dre_uuid,
+        periodo_uuid=periodo_uuid,
+        usuario=usuario,
+        parcial=parcial,
+        consolidado_dre_uuid=consolidado_dre_uuid,
+        sequencia_de_publicacao=sequencia_de_publicacao,
+        apenas_nao_publicadas=apenas_nao_publicadas,
+    )
+
+    verificar_se_gera_ata_parecer_tecnico(
+        dre=dre,
+        periodo=periodo,
+        consolidado_dre=consolidado_dre,
+        usuario=usuario,
+        ata=ata,
+        parcial=parcial,
+    )
+
+    gerar_lauda_txt_consolidado_dre(
+        consolidado_dre=consolidado_dre,
+        dre=dre,
+        periodo=periodo,
+        parcial=parcial,
+        usuario=usuario,
+        apenas_nao_publicadas=apenas_nao_publicadas,
+    )
+
+    passar_pcs_do_relatorio_para_publicadas(
+        dre=dre,
+        periodo=periodo,
+        consolidado_dre=consolidado_dre
+    )
+
+    eh_parcial = parcial['parcial']
+    consolidado_dre.passar_para_status_gerado(eh_parcial)
+
+    if (eh_parcial):
+        verifica_se_relatorio_consolidado_deve_ser_gerado(
+            dre=dre,
+            periodo=periodo,
+            usuario=usuario
+        )
